FIR/UAST: commonize UTryExpression and UCatchClause
