+++ uast/uast-kotlin-base/src/org/jetbrains/uast/kotlin/controlStructures/KotlinUCatchClause.kt
@@ -1,41 +1,26 @@
-/*
- * Copyright 2010-2016 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2000-2021 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 
 package org.jetbrains.uast.kotlin
 
 import com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.psi.KtCatchClause
-import org.jetbrains.uast.UCatchClause
-import org.jetbrains.uast.UElement
-import org.jetbrains.uast.UParameter
-import org.jetbrains.uast.UTypeReferenceExpression
+import org.jetbrains.uast.*
 import org.jetbrains.uast.internal.acceptList
 import org.jetbrains.uast.kotlin.psi.UastKotlinPsiParameter
 import org.jetbrains.uast.visitor.UastVisitor
 
 class KotlinUCatchClause(
-        override val sourcePsi: KtCatchClause,
-        givenParent: UElement?
+    override val sourcePsi: KtCatchClause,
+    givenParent: UElement?
 ) : KotlinAbstractUElement(givenParent), UCatchClause {
-    override val psi: PsiElement?
+    override val psi: PsiElement
         get() = sourcePsi
 
     override val javaPsi: PsiElement? get() = null
 
-    override val body by lz { KotlinConverter.convertOrEmpty(sourcePsi.catchBody, this) }
+    override val body by lz {
+        baseResolveProviderService.baseKotlinConverter.convertOrEmpty(sourcePsi.catchBody, this)
+    }
 
     override val parameters by lz {
         val parameter = sourcePsi.catchParameter ?: return@lz emptyList<UParameter>()
@@ -47,7 +32,11 @@ class KotlinUCatchClause(
     override val typeReferences by lz {
         val parameter = sourcePsi.catchParameter ?: return@lz emptyList<UTypeReferenceExpression>()
         val typeReference = parameter.typeReference ?: return@lz emptyList<UTypeReferenceExpression>()
-        listOf(KotlinUTypeReferenceExpression(typeReference, this, baseResolveProviderService) { typeReference.toPsiType(this, boxed = true) })
+        listOf(
+            KotlinUTypeReferenceExpression(typeReference, this, baseResolveProviderService) {
+                baseResolveProviderService.resolveToType(typeReference, this) ?: UastErrorType
+            }
+        )
     }
 
     // equal to IDEA 202 implementation
