+++ uast/uast-kotlin-fir/src/org/jetbrains/uast/kotlin/FirKotlinConverter.kt
@@ -242,6 +242,9 @@ internal object FirKotlinConverter : BaseKotlinConverter {
                 is KtClassBody -> {
                     el<UExpressionList>(build(KotlinUExpressionList.Companion::createClassBody))
                 }
+                is KtCatchClause -> {
+                    el<UCatchClause>(build(::KotlinUCatchClause))
+                }
                 is KtVariableDeclaration -> {
                     if (element is KtProperty && !element.isLocal) {
                         convertNonLocalProperty(element, givenParent, this).firstOrNull()
@@ -369,6 +372,7 @@ internal object FirKotlinConverter : BaseKotlinConverter {
                 }
                 is KtReturnExpression -> expr<UReturnExpression>(build(::KotlinUReturnExpression))
                 is KtThrowExpression -> expr<UThrowExpression>(build(::KotlinUThrowExpression))
+                is KtTryExpression -> expr<UTryExpression>(build(::KotlinUTryExpression))
 
                 is KtBreakExpression -> expr<UBreakExpression>(build(::KotlinUBreakExpression))
                 is KtContinueExpression -> expr<UContinueExpression>(build(::KotlinUContinueExpression))
