+++ uast/uast-kotlin-fir/testData/legacyIdentifiers/TryCatch.identifiers.fir.txt
@@ -1,7 +1,7 @@
 TryCatch -> UClass (name = TryCatch)
 catches -> UMethod (name = catches)
         body -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
-    e -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
+    e -> UParameter (name = e)
     Throwable -> USimpleNameReferenceExpression (identifier = Throwable)
         catcher -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
         finalizer -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
@@ -17,7 +17,7 @@ catches -> UMethod (name = catches)
     java -> USimpleNameReferenceExpression (identifier = java)
     lang -> USimpleNameReferenceExpression (identifier = lang)
     SuppressWarnings -> USimpleNameReferenceExpression (identifier = SuppressWarnings)
-    e -> [!] UnknownKotlinExpression (VALUE_PARAMETER)
+    e -> UParameter (name = e)
     Throwable -> USimpleNameReferenceExpression (identifier = Throwable)
         catcher -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
         finalizer -> [!] UnknownKotlinExpression (CALL_EXPRESSION)
